move MetaInfo to a seperate file, to remove dependency on utils where possible
