+++ srcs/chattisgarh.py
@@ -7,6 +7,7 @@ import datetime
 
 from ..utils  import utils
 from .central import CentralBase 
+from ..utils.metainfo import MetaInfo
 
 class ChattisgarhWeekly(CentralBase):
     def __init__(self, name, storage):
@@ -86,7 +87,7 @@ class ChattisgarhWeekly(CentralBase):
                     gznum = txt
                 elif col.startswith('partnum'):
                     h, partnum = col.split('|')
-                    metainfo = utils.MetaInfo()
+                    metainfo = MetaInfo()
                     metainfos.append(metainfo)
                     metainfo.set_date(dateobj)
                     metainfo.set_gztype(self.gazette_type)
@@ -229,7 +230,7 @@ class ChattisgarhExtraordinary(ChattisgarhWeekly):
         return order
 
     def process_result_row(self, tr, metainfos, dateobj, order):
-        metainfo = utils.MetaInfo()
+        metainfo = MetaInfo()
         metainfo.set_gztype(self.gazette_type)
         metainfos.append(metainfo)
         metainfo.set_date(dateobj)
