+++ srcs/goa.py
@@ -4,6 +4,7 @@ import os
 
 from .basegazette import BaseGazette
 from ..utils import utils
+from ..utils.metainfo import MetaInfo
 
 class Goa(BaseGazette):
     def __init__(self, name, storage):
@@ -91,7 +92,7 @@ class Goa(BaseGazette):
         return minfos
 
     def get_metainfo(self, order, tr, dateobj):
-        metainfo = utils.MetaInfo()
+        metainfo = MetaInfo()
         metainfo.set_date(dateobj)
 
         i = 0
