+++ srcs/karnataka.py
@@ -6,6 +6,7 @@ import datetime
 
 from ..utils import utils
 from .basegazette import BaseGazette
+from ..utils.metainfo import MetaInfo
 
 class Karnataka(BaseGazette):
     def __init__(self, name, storage):
@@ -31,7 +32,7 @@ class Karnataka(BaseGazette):
         dateurl = self.baseurl % datestr
         docurl  = urllib.parse.urljoin(dateurl, mainhref)
 
-        mainmeta = utils.MetaInfo()
+        mainmeta = MetaInfo()
         mainmeta.set_date(dateobj)
         mainmeta.set_url(self.url_fix(docurl))
        
@@ -67,7 +68,7 @@ class Karnataka(BaseGazette):
             relurl = os.path.join(relpath, partnum)
             docurl = urllib.parse.urljoin(dateurl, href) 
 
-            metainfo = utils.MetaInfo()
+            metainfo = MetaInfo()
             metainfo.set_date(dateobj)
             metainfo['partnum'] = partnum
