+++ srcs/kerala.py
@@ -5,6 +5,7 @@ import urllib.request, urllib.parse, urllib.error
 
 from .basegazette import BaseGazette
 from ..utils import utils
+from ..utils.metainfo import MetaInfo
 
 class Kerala(BaseGazette):
     def __init__(self, name, storage):
@@ -163,7 +164,7 @@ class Kerala(BaseGazette):
                         title = utils.get_tag_contents(x)
                         title = utils.remove_spaces(title)
 
-                        metainfo = utils.MetaInfo()
+                        metainfo = MetaInfo()
                         minfos.append(metainfo)
 
                         metainfo.set_title(title)
