+++ srcs/madhyapradesh.py
@@ -5,6 +5,7 @@ import urllib.request, urllib.parse, urllib.error
 
 from .basegazette import BaseGazette
 from ..utils import utils
+from ..utils.metainfo import MetaInfo
 
 class MadhyaPradesh(BaseGazette):
     def __init__(self, name, storage):
@@ -94,7 +95,7 @@ class MadhyaPradesh(BaseGazette):
                 if date != dateobj:
                     continue
                                 
-                metainfo = utils.MetaInfo()
+                metainfo = MetaInfo()
                 metainfo.set_gztype(gztype)
                 metainfo.set_date(dateobj)
                 metainfo.set_url(url)
@@ -169,7 +170,7 @@ class MadhyaPradesh(BaseGazette):
         return None     
     
     def process_row(self, tr, order, dateobj):    
-        metainfo = utils.MetaInfo()
+        metainfo = MetaInfo()
         i = 0
         for td in tr.find_all('td'):
             txt = utils.get_tag_contents(td)
