+++ utils/metainfo.py
@@ -1,0 +1,78 @@
+import re
+URL        = 'url'
+HREF       = 'href'
+TITLE      = 'title'
+DATE       = 'date'
+MINISTRY   = 'ministry'
+SUBJECT    = 'subject'
+GZTYPE     = 'gztype'
+
+_illegal_xml_chars_RE = re.compile('[\x00-\x08\x0b\x0c\x0e-\x1F\uD800-\uDFFF\uFFFE\uFFFF]')
+
+def replace_xml_illegal_chars(val, replacement=' '):
+    """Filter out characters that are illegal in XML."""
+
+    return _illegal_xml_chars_RE.sub(replacement, val)
+
+class MetaInfo(dict):
+    def __init__(self):
+        dict.__init__(self)
+
+    def copy(self):
+        m = MetaInfo()
+        for k, v in self.items():
+            m[k] = v
+        return m
+ 
+    def set_field(self, field, value):
+        if type(value) in (str,):
+            value = replace_xml_illegal_chars(value)
+        self.__setitem__(field, value)
+
+    def get_field(self, field):
+        if field in self:
+            return self.get(field)
+        return None
+
+    def set_date(self, value):
+        self.set_field(DATE, value)
+
+    def set_title(self, value):
+        self.set_field(TITLE, value)
+
+    def set_url(self, value):
+        self.set_field(URL, value)
+
+    def set_href(self, value):
+        self.set_field(HREF, value)
+
+    def set_subject(self, value):
+        self.set_field(SUBJECT, value)
+
+    def set_ministry(self, value):
+        self.set_field(MINISTRY, value)
+
+    def set_gztype(self, value):
+        self.set_field(GZTYPE, value)
+
+    def get_url(self):
+        return self.get_field(URL)
+
+    def get_href(self):
+        return self.get_field(HREF)
+
+    def get_title(self):
+        return self.get_field(TITLE)
+
+    def get_date(self):
+        return self.get_field(DATE)
+
+    def get_ministry(self):
+        return self.get_field(MINISTRY)
+
+    def get_subject(self):
+        return self.get_field(SUBJECT)
+
+    def get_gztype(self, value):
+        return self.get_field(GZTYPE)
+
